Moved types/ dir
